Fix lang setting not loading.
